refactor(coordinator/store): Phase 1 — split Store interface into per-domain fragments
Declarations-only. The 1049-line interface.go is split into per-domain files,
each holding its record/DTO types plus an interface fragment; interface.go now
composes them by embedding:

  type Store interface { APIKeyStore; UsageStore; LedgerStore; ... }

New files: apikeys, usage, ledger, referrals, billing_sessions, pricing,
model_registry, releases, users, stripe_withdrawals, device_codes, invites,
provider_earnings, provider_tokens, providers, reputation, log_reports,
telemetry. ProviderLocation now has a single home (providers.go);
ErrInsufficientBalance stays a package var (ledger.go).

Zero method-set change: proven by `var _ Store` in both impls plus a
whole-module compile of every caller. All 313 JSON struct tags verified
byte-identical. Build, vet, gofmt, go test ./coordinator/store/... green.
