+++ coordinator/store/billing_sessions.go
@@ -1,0 +1,32 @@
+package store
+
+import "time"
+
+// BillingSessionStore covers in-progress deposit sessions (Stripe checkout).
+type BillingSessionStore interface {
+	// CreateBillingSession stores a new billing session (Stripe).
+	CreateBillingSession(session *BillingSession) error
+
+	// GetBillingSession retrieves a billing session by ID.
+	GetBillingSession(sessionID string) (*BillingSession, error)
+
+	// CompleteBillingSession marks a session as completed and sets the completion time.
+	CompleteBillingSession(sessionID string) error
+
+	// IsExternalIDProcessed returns true if a billing session with this external ID
+	// has already been completed. Used to prevent double-crediting the same on-chain tx.
+	IsExternalIDProcessed(externalID string) bool
+}
+
+// BillingSession tracks an in-progress payment via any method (Stripe).
+type BillingSession struct {
+	ID             string     `json:"id"`
+	AccountID      string     `json:"account_id"`
+	PaymentMethod  string     `json:"payment_method"` // "stripe"
+	AmountMicroUSD int64      `json:"amount_micro_usd"`
+	ExternalID     string     `json:"external_id"`   // Stripe session ID, tx hash, etc.
+	Status         string     `json:"status"`        // "pending", "completed", "expired"
+	ReferralCode   string     `json:"referral_code"` // optional
+	CreatedAt      time.Time  `json:"created_at"`
+	CompletedAt    *time.Time `json:"completed_at,omitempty"`
+}
