+++ coordinator/store/referrals.go
@@ -1,0 +1,39 @@
+package store
+
+import "time"
+
+// ReferralStore covers the referral system (partners earn a share of the
+// platform fee from accounts they refer).
+type ReferralStore interface {
+	// CreateReferrer registers an account as a referrer with the given code.
+	CreateReferrer(accountID, code string) error
+
+	// GetReferrerByCode returns the referrer for a given referral code.
+	GetReferrerByCode(code string) (*Referrer, error)
+
+	// GetReferrerByAccount returns the referrer record for an account, if registered.
+	GetReferrerByAccount(accountID string) (*Referrer, error)
+
+	// RecordReferral records that referredAccountID was referred by referrerCode.
+	RecordReferral(referrerCode, referredAccountID string) error
+
+	// GetReferrerForAccount returns the referrer code that referred this account, or "" if none.
+	GetReferrerForAccount(accountID string) (string, error)
+
+	// GetReferralStats returns referral statistics for a code.
+	GetReferralStats(code string) (*ReferralStats, error)
+}
+
+// Referrer represents a registered referral partner.
+type Referrer struct {
+	AccountID string    `json:"account_id"`
+	Code      string    `json:"code"`
+	CreatedAt time.Time `json:"created_at"`
+}
+
+// ReferralStats provides aggregate metrics for a referral code.
+type ReferralStats struct {
+	Code                 string `json:"code"`
+	TotalReferred        int    `json:"total_referred"`
+	TotalRewardsMicroUSD int64  `json:"total_rewards_micro_usd"`
+}
